Plugin SystemInformation ====================== - Correct the MemoryLoad keyword for Linux. - Correct RAM and Virtual Process Memory (divide by 1000) for Windows. - Change historizer Byte and KByte from long long to long.
Plugin WeatherUnderground
========================
- Correct some syntaxes.
- Correct visual unit in debug mode for keyword Temp and WindDirection.

Widget Forecast
==============
- Correct the symbol degree °, to be correctly display under Linux.

+++ trunk/sources/plugins/SystemInformation/specificHistorizers/Byte.cpp
@@ -9,7 +9,7 @@ namespace shared { namespace plugin { namespace yPluginApi { namespace historiza
    const CStandardCapacity& ByteCapacity = CStandardCapacity("Byte", ByteUnit, EKeywordDataType::kNumeric);
 
    CByte::CByte(const std::string& keywordName, const EKeywordAccessMode& accessMode, const EMeasureType& measureType, typeInfo::CIntTypeInfo & additionalInfo)
-      :CSingleHistorizableData<long long>(keywordName, ByteCapacity, accessMode, 0, measureType, additionalInfo)
+      :CSingleHistorizableData<long>(keywordName, ByteCapacity, accessMode, 0, measureType, additionalInfo)
    {}
 
    CByte::~CByte()
